fix: HMAC for JS platform
- SHA-224
- SHA-256
- SHA-384
- SHA-512
