+++ Hashing/src/commonMain/kotlin/io/iohk/atala/prism/apollo/hashing/internal/HMACInterface.kt
@@ -9,7 +9,7 @@ interface HMACInterface {
     public fun createHmac(key: ByteArray, outputLength: Int? = null): Digest = HMAC(this as Digest, key, outputLength)
 
     /**
-     * Create an [HMAC] hash of [input] using the the provided algorithm.
+     * Create an [HMAC] hash of [input] using the provided algorithm.
      */
     public fun hmac(key: ByteArray, input: ByteArray, outputLength: Int? = null): ByteArray = createHmac(key, outputLength).digest(input)
 }
